feat(qa): Q&A検索結果表示コンポーネントの追加と検索機能のリファクタリング
